Initialize the keyboard at JCFB startup
We use ncurses to handle keyboard in a gentle way. Indeed, doing all from
scratch is pretty difficult, and not as fun as that. Ncurses is a piece of
shit, but it does the trick.

Until now, the keyboard system is simple: when the user call "jcfb_refresh",
JCFB read every keys pressed since last refresh, and add every key press in
a polling queue. The user can then use "jcfb_next_key" to get every pressed
key in a chronologic order.

In the future, we will maintain a state for every keys allowing to say
whereas a key is pressed, repeated or released.