Bitmap optimisation
This commit introduces an optimisation when manipulating bitmap.

Bitmaps' pixels are now coded with a fixed size. Pixels are now not packed
in memory, allowing to use memcpy on bitmaps of different pixel formats.
Rendering process are faster using this way (memory is aligned, copy from
bitmaps to other is trival).

The internal framebuffer memory being packed, we need to pack buffer bitmap
when drawing it to the screen. This is done with a naive way, but we could
improve this process using SSE. However, it is not clear now whereas this is
a critical code pass. I guess not, since major rendering will be done on
bitmaps. Screen refresh is happening only once per frame.