Better support of mouse speed
We introduced an API to get immediate mouse speed.

The speed was not correctly computed since we gave the mouse speed as
the speed of the last event handled by `mouse_update()`.

To correct that, we instead define the mouse speed as the difference
between the mouse position before the `mouse_update()` and after the
`mouse_update()`.