Start alpha-channel support
Previously, we used the bright-purple color as masked color, and we
converted pixels with alpha component to bright purple in
`bitmap_load()`.

Now, we conserve the alpha channel if the framebuffer pixel format has
one. In case it is RGB24, we conserve it too, in the last byte.

Scaling functions do mask check, this has no great impact on
performances.