Macro-parametrized bitmap blit functions
Bitmap blit functions will in the future be able to work in many other ways
that simply provide pixel copy operation. For exemple, one want to have
blitting using additive blending operation.

To ease this move, we provide a "macro-parametrized" blit functions file in
src/bitmap-blit.inc.c

This file is included in the bitmap.c file, and is parametrized with the
`BLIT_PIXEL_FUNC(dst, src)` macro, that corresponds to the pixel write
operation from `src` to `dst` and the `BLIT_FUNC_SUFFIX` define, that add
a suffix to every blit functions.